+++ src/main/java/tss/g2/fyre/controllers/javalin/AuthorizationController.java
@@ -4,6 +4,7 @@ import io.javalin.Javalin;
 import tss.g2.fyre.controllers.CreateController;
 import tss.g2.fyre.models.Answer;
 import tss.g2.fyre.models.actions.CheckAuthorization;
+import tss.g2.fyre.models.actions.RegisterUser;
 import tss.g2.fyre.models.datastorage.DataStorage;
 
 /**
@@ -31,6 +32,18 @@ public class AuthorizationController implements CreateController {
       String login = ctx.queryParam("login");
       String password = ctx.queryParam("password");
       Answer answer = new CheckAuthorization(dataStorage, login, password).getAnswer();
+      ctx.result(answer.toJson());
+    });
+
+    app.post("/registration", ctx -> {
+      String login = ctx.queryParam("login");
+      String password = ctx.queryParam("password");
+      String name = ctx.queryParam("name");
+      String surname = ctx.queryParam("surname");
+      String email = ctx.queryParam("email");
+
+      Answer answer = new RegisterUser(dataStorage, login, password, name, surname, email)
+              .getAnswer();
       ctx.result(answer.toJson());
     });
   }
