+++ src/main/java/tss/g2/fyre/models/actions/RegisterUser.java
@@ -1,0 +1,39 @@
+package tss.g2.fyre.models.actions;
+
+import tss.g2.fyre.models.Answer;
+import tss.g2.fyre.models.datastorage.DataStorage;
+
+public class RegisterUser implements Action {
+  private DataStorage dataStorage;
+  private String login;
+  private String password;
+  private String name;
+  private String surname;
+  private String email;
+
+  /**
+   * Constructor.
+   *
+   * @param dataStorage data storage object
+   * @param login user login
+   * @param password user password
+   * @param name user name
+   * @param surname user surname
+   * @param email user email
+   */
+  public RegisterUser(DataStorage dataStorage, String login, String password,
+                      String name, String surname, String email) {
+    this.dataStorage = dataStorage;
+    this.login = login;
+    this.password = password;
+    this.name = name;
+    this.surname = surname;
+    this.email = email;
+  }
+
+  @Override
+  public Answer getAnswer() {
+    return new Answer<>(true,
+            dataStorage.createNewPerson(login, password, name, surname, email));
+  }
+}
